% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_animal_age.R
\name{convert_animal_age}
\alias{convert_animal_age}
\title{Exclaim}
\usage{
convert_animal_age(animal, human_age)
}
\arguments{
\item{animal}{dog or goat}

\item{human_age}{human age equivalent for animal}
}
\description{
Exclaim
}
\examples{
convert_animal_age(animal = goat, human_age = 12)
}
