% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backpack_weight.R
\name{backpack_weight}
\alias{backpack_weight}
\title{Calculate backpack weight based on weight of items inside.}
\usage{
backpack_weight(pack, lunch, laptop, jacket)
}
\arguments{
\item{pack}{weight of empty pack in g}

\item{lunch}{weight of lunchbox in g}

\item{laptop}{weight of laptop in g}

\item{jacket}{weight of jacket in g}
}
\value{
weight of entire backpack in g
}
\description{
Calculate backpack weight based on weight of items inside.
}
\examples{
backpack_weight(pack = 5000, lunch = 2000, laptop = 1120, jacket = 1050)
}
